requirements.txt
streamlit
opencv-python
pyzbar
pillow
numpy
matplotlib
PaddleOCR
paddlepaddle
joblib
pytesseract
requests
sqlite3
sklearn
seaborn
xgboost
urlib.parse
flask
traceback
tldextract
datetime
time
uuid